[PM-3175] Replaced OrganizationService.InviteUserAsync with command
