+++ src/Core/OrganizationFeatures/OrganizationUsers/UpdateOrganizationUserCommand.cs
@@ -53,7 +53,7 @@ public class UpdateOrganizationUserCommand : OrganizationUserCommand, IUpdateOrg
 
         if (savingUserId.HasValue)
         {
-            await ValidateOrganizationUserUpdatePermissions(user.OrganizationId, user.Type, originalUser.Type, user.GetPermissions());
+            await ValidateOrganizationUserUpdatePermissionsAsync(user.OrganizationId, user.Type, originalUser.Type, user.GetPermissions());
         }
 
         await ValidateOrganizationCustomPermissionsEnabledAsync(user.OrganizationId, user.Type);
