+++ test/Core.Test/Services/OrganizationServiceTests.cs
@@ -1,7 +1,5 @@
-﻿using System.Text.Json;
-using Bit.Core.Auth.Entities;
+﻿using Bit.Core.Auth.Entities;
 using Bit.Core.Auth.Enums;
-using Bit.Core.Auth.Models.Business;
 using Bit.Core.Auth.Models.Data;
 using Bit.Core.Auth.Repositories;
 using Bit.Core.Context;
@@ -11,10 +9,7 @@ using Bit.Core.Enums;
 using Bit.Core.Enums.Provider;
 using Bit.Core.Exceptions;
 using Bit.Core.Models.Business;
-using Bit.Core.Models.Data;
 using Bit.Core.Models.StaticStore;
-using Bit.Core.OrganizationFeatures.OrganizationSubscriptions.Interface;
-using Bit.Core.OrganizationFeatures.OrganizationUsers.Interfaces;
 using Bit.Core.Repositories;
 using Bit.Core.Services;
 using Bit.Core.Settings;
@@ -28,7 +23,6 @@ using Bit.Core.Utilities;
 using Bit.Test.Common.AutoFixture;
 using Bit.Test.Common.AutoFixture.Attributes;
 using NSubstitute;
-using NSubstitute.ExceptionExtensions;
 using Xunit;
 using Organization = Bit.Core.Entities.Organization;
 using OrganizationUser = Bit.Core.Entities.OrganizationUser;
@@ -149,425 +143,6 @@ public class OrganizationServiceTests
         var exception = await Assert.ThrowsAsync<BadRequestException>(
             () => sutProvider.Sut.SignUpAsync(signup));
         Assert.Contains("You can't subtract Service Accounts!", exception.Message);
-    }
-
-    [Theory]
-    [OrganizationInviteCustomize(InviteeUserType = OrganizationUserType.User,
-         InvitorUserType = OrganizationUserType.Owner), BitAutoData]
-    public async Task InviteUser_NoEmails_Throws(Organization organization, OrganizationUser invitor,
-        OrganizationUserInvite invite, SutProvider<OrganizationService> sutProvider)
-    {
-        invite.Emails = null;
-        sutProvider.GetDependency<ICurrentContext>().OrganizationOwner(organization.Id).Returns(true);
-        sutProvider.GetDependency<ICurrentContext>().ManageUsers(organization.Id).Returns(true);
-        sutProvider.GetDependency<IOrganizationRepository>().GetByIdAsync(organization.Id).Returns(organization);
-        await Assert.ThrowsAsync<NotFoundException>(
-            () => sutProvider.Sut.InviteUsersAsync(organization.Id, invitor.UserId, new (OrganizationUserInvite, string)[] { (invite, null) }));
-    }
-
-    [Theory]
-    [OrganizationInviteCustomize, BitAutoData]
-    public async Task InviteUser_DuplicateEmails_PassesWithoutDuplicates(Organization organization, OrganizationUser invitor,
-                [OrganizationUser(OrganizationUserStatusType.Confirmed, OrganizationUserType.Owner)] OrganizationUser owner,
-        OrganizationUserInvite invite, SutProvider<OrganizationService> sutProvider)
-    {
-        invite.Emails = invite.Emails.Append(invite.Emails.First());
-
-        sutProvider.GetDependency<IOrganizationRepository>().GetByIdAsync(organization.Id).Returns(organization);
-        sutProvider.GetDependency<ICurrentContext>().OrganizationOwner(organization.Id).Returns(true);
-        sutProvider.GetDependency<ICurrentContext>().ManageUsers(organization.Id).Returns(true);
-        var organizationUserRepository = sutProvider.GetDependency<IOrganizationUserRepository>();
-        organizationUserRepository.GetManyByOrganizationAsync(organization.Id, OrganizationUserType.Owner)
-            .Returns(new[] { owner });
-
-        await sutProvider.Sut.InviteUsersAsync(organization.Id, invitor.UserId, new (OrganizationUserInvite, string)[] { (invite, null) });
-
-        await sutProvider.GetDependency<IMailService>().Received(1)
-            .BulkSendOrganizationInviteEmailAsync(organization.Name,
-                Arg.Is<IEnumerable<(OrganizationUser, ExpiringToken)>>(v => v.Count() == invite.Emails.Distinct().Count()), organization.PlanType == PlanType.Free);
-    }
-
-    [Theory]
-    [OrganizationInviteCustomize(
-        InviteeUserType = OrganizationUserType.Admin,
-        InvitorUserType = OrganizationUserType.Owner
-    ), BitAutoData]
-    public async Task InviteUser_NoOwner_Throws(Organization organization, OrganizationUser invitor,
-        OrganizationUserInvite invite, SutProvider<OrganizationService> sutProvider)
-    {
-        sutProvider.GetDependency<IOrganizationRepository>().GetByIdAsync(organization.Id).Returns(organization);
-        sutProvider.GetDependency<ICurrentContext>().OrganizationOwner(organization.Id).Returns(true);
-        sutProvider.GetDependency<ICurrentContext>().ManageUsers(organization.Id).Returns(true);
-        var exception = await Assert.ThrowsAsync<BadRequestException>(
-            () => sutProvider.Sut.InviteUsersAsync(organization.Id, invitor.UserId, new (OrganizationUserInvite, string)[] { (invite, null) }));
-        Assert.Contains("Organization must have at least one confirmed owner.", exception.Message);
-    }
-
-    [Theory]
-    [OrganizationInviteCustomize(
-        InviteeUserType = OrganizationUserType.Owner,
-        InvitorUserType = OrganizationUserType.Admin
-    ), BitAutoData]
-    public async Task InviteUser_NonOwnerConfiguringOwner_Throws(Organization organization, OrganizationUserInvite invite,
-        OrganizationUser invitor, SutProvider<OrganizationService> sutProvider)
-    {
-        var organizationRepository = sutProvider.GetDependency<IOrganizationRepository>();
-        var currentContext = sutProvider.GetDependency<ICurrentContext>();
-
-        organizationRepository.GetByIdAsync(organization.Id).Returns(organization);
-        currentContext.OrganizationAdmin(organization.Id).Returns(true);
-
-        var exception = await Assert.ThrowsAsync<BadRequestException>(
-            () => sutProvider.Sut.InviteUsersAsync(organization.Id, invitor.UserId, new (OrganizationUserInvite, string)[] { (invite, null) }));
-        Assert.Contains("only an owner", exception.Message.ToLowerInvariant());
-    }
-
-    [Theory]
-    [OrganizationInviteCustomize(
-        InviteeUserType = OrganizationUserType.Custom,
-        InvitorUserType = OrganizationUserType.User
-    ), BitAutoData]
-    public async Task InviteUser_NonAdminConfiguringAdmin_Throws(Organization organization, OrganizationUserInvite invite,
-        OrganizationUser invitor, SutProvider<OrganizationService> sutProvider)
-    {
-        organization.UseCustomPermissions = true;
-
-        var organizationRepository = sutProvider.GetDependency<IOrganizationRepository>();
-        var currentContext = sutProvider.GetDependency<ICurrentContext>();
-
-        organizationRepository.GetByIdAsync(organization.Id).Returns(organization);
-        currentContext.OrganizationUser(organization.Id).Returns(true);
-
-        var exception = await Assert.ThrowsAsync<BadRequestException>(
-            () => sutProvider.Sut.InviteUsersAsync(organization.Id, invitor.UserId, new (OrganizationUserInvite, string)[] { (invite, null) }));
-        Assert.Contains("your account does not have permission to manage users", exception.Message.ToLowerInvariant());
-    }
-
-    [Theory]
-    [OrganizationInviteCustomize(
-         InviteeUserType = OrganizationUserType.Custom,
-         InvitorUserType = OrganizationUserType.Admin
-     ), BitAutoData]
-    public async Task InviteUser_WithCustomType_WhenUseCustomPermissionsIsFalse_Throws(Organization organization, OrganizationUserInvite invite,
-        OrganizationUser invitor, SutProvider<OrganizationService> sutProvider)
-    {
-        organization.UseCustomPermissions = false;
-
-        invite.Permissions = null;
-        invitor.Status = OrganizationUserStatusType.Confirmed;
-        var organizationRepository = sutProvider.GetDependency<IOrganizationRepository>();
-        var organizationUserRepository = sutProvider.GetDependency<IOrganizationUserRepository>();
-        var currentContext = sutProvider.GetDependency<ICurrentContext>();
-
-        organizationRepository.GetByIdAsync(organization.Id).Returns(organization);
-        organizationUserRepository.GetManyByOrganizationAsync(organization.Id, OrganizationUserType.Owner)
-            .Returns(new[] { invitor });
-        currentContext.OrganizationOwner(organization.Id).Returns(true);
-        currentContext.ManageUsers(organization.Id).Returns(true);
-
-        var exception = await Assert.ThrowsAsync<BadRequestException>(
-            () => sutProvider.Sut.InviteUsersAsync(organization.Id, invitor.UserId, new (OrganizationUserInvite, string)[] { (invite, null) }));
-        Assert.Contains("to enable custom permissions", exception.Message.ToLowerInvariant());
-    }
-
-    [Theory]
-    [OrganizationInviteCustomize(
-         InviteeUserType = OrganizationUserType.Custom,
-         InvitorUserType = OrganizationUserType.Admin
-     ), BitAutoData]
-    public async Task InviteUser_WithCustomType_WhenUseCustomPermissionsIsTrue_Passes(Organization organization, OrganizationUserInvite invite,
-        OrganizationUser invitor, SutProvider<OrganizationService> sutProvider)
-    {
-        organization.Seats = 10;
-        organization.UseCustomPermissions = true;
-
-        invite.Permissions = null;
-        invitor.Status = OrganizationUserStatusType.Confirmed;
-        var organizationRepository = sutProvider.GetDependency<IOrganizationRepository>();
-        var organizationUserRepository = sutProvider.GetDependency<IOrganizationUserRepository>();
-        var currentContext = sutProvider.GetDependency<ICurrentContext>();
-
-        organizationRepository.GetByIdAsync(organization.Id).Returns(organization);
-        organizationUserRepository.GetManyByOrganizationAsync(organization.Id, OrganizationUserType.Owner)
-            .Returns(new[] { invitor });
-        currentContext.OrganizationOwner(organization.Id).Returns(true);
-        currentContext.ManageUsers(organization.Id).Returns(true);
-
-        await sutProvider.Sut.InviteUsersAsync(organization.Id, invitor.UserId, new (OrganizationUserInvite, string)[] { (invite, null) });
-    }
-
-    [Theory]
-    [BitAutoData(OrganizationUserType.Admin)]
-    [BitAutoData(OrganizationUserType.Manager)]
-    [BitAutoData(OrganizationUserType.Owner)]
-    [BitAutoData(OrganizationUserType.User)]
-    public async Task InviteUser_WithNonCustomType_WhenUseCustomPermissionsIsFalse_Passes(OrganizationUserType inviteUserType, Organization organization, OrganizationUserInvite invite,
-        OrganizationUser invitor, SutProvider<OrganizationService> sutProvider)
-    {
-        organization.Seats = 10;
-        organization.UseCustomPermissions = false;
-
-        invite.Type = inviteUserType;
-        invite.Permissions = null;
-        invitor.Status = OrganizationUserStatusType.Confirmed;
-        var organizationRepository = sutProvider.GetDependency<IOrganizationRepository>();
-        var organizationUserRepository = sutProvider.GetDependency<IOrganizationUserRepository>();
-        var currentContext = sutProvider.GetDependency<ICurrentContext>();
-
-        organizationRepository.GetByIdAsync(organization.Id).Returns(organization);
-        organizationUserRepository.GetManyByOrganizationAsync(organization.Id, OrganizationUserType.Owner)
-            .Returns(new[] { invitor });
-        currentContext.OrganizationOwner(organization.Id).Returns(true);
-        currentContext.ManageUsers(organization.Id).Returns(true);
-
-        await sutProvider.Sut.InviteUsersAsync(organization.Id, invitor.UserId, new (OrganizationUserInvite, string)[] { (invite, null) });
-    }
-
-    [Theory]
-    [OrganizationInviteCustomize(
-        InviteeUserType = OrganizationUserType.Manager,
-        InvitorUserType = OrganizationUserType.Custom
-    ), BitAutoData]
-    public async Task InviteUser_CustomUserWithoutManageUsersConfiguringUser_Throws(Organization organization, OrganizationUserInvite invite,
-        OrganizationUser invitor, SutProvider<OrganizationService> sutProvider)
-    {
-        invitor.Permissions = JsonSerializer.Serialize(new Permissions() { ManageUsers = false },
-            new JsonSerializerOptions
-            {
-                PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
-            });
-
-        var organizationRepository = sutProvider.GetDependency<IOrganizationRepository>();
-        var currentContext = sutProvider.GetDependency<ICurrentContext>();
-
-        organizationRepository.GetByIdAsync(organization.Id).Returns(organization);
-        currentContext.OrganizationCustom(organization.Id).Returns(true);
-        currentContext.ManageUsers(organization.Id).Returns(false);
-
-        var exception = await Assert.ThrowsAsync<BadRequestException>(
-            () => sutProvider.Sut.InviteUsersAsync(organization.Id, invitor.UserId, new (OrganizationUserInvite, string)[] { (invite, null) }));
-        Assert.Contains("account does not have permission", exception.Message.ToLowerInvariant());
-    }
-
-    [Theory]
-    [OrganizationInviteCustomize(
-        InviteeUserType = OrganizationUserType.Admin,
-        InvitorUserType = OrganizationUserType.Custom
-    ), BitAutoData]
-    public async Task InviteUser_CustomUserConfiguringAdmin_Throws(Organization organization, OrganizationUserInvite invite,
-        OrganizationUser invitor, SutProvider<OrganizationService> sutProvider)
-    {
-        invitor.Permissions = JsonSerializer.Serialize(new Permissions() { ManageUsers = true },
-            new JsonSerializerOptions
-            {
-                PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
-            });
-
-        var organizationRepository = sutProvider.GetDependency<IOrganizationRepository>();
-        var currentContext = sutProvider.GetDependency<ICurrentContext>();
-
-        organizationRepository.GetByIdAsync(organization.Id).Returns(organization);
-        currentContext.OrganizationCustom(organization.Id).Returns(true);
-        currentContext.ManageUsers(organization.Id).Returns(true);
-
-        var exception = await Assert.ThrowsAsync<BadRequestException>(
-            () => sutProvider.Sut.InviteUsersAsync(organization.Id, invitor.UserId, new (OrganizationUserInvite, string)[] { (invite, null) }));
-        Assert.Contains("can not manage admins", exception.Message.ToLowerInvariant());
-    }
-
-    [Theory]
-    [OrganizationInviteCustomize(
-        InviteeUserType = OrganizationUserType.User,
-        InvitorUserType = OrganizationUserType.Owner
-    ), BitAutoData]
-    public async Task InviteUser_NoPermissionsObject_Passes(Organization organization, OrganizationUserInvite invite,
-        OrganizationUser invitor, SutProvider<OrganizationService> sutProvider)
-    {
-        invite.Permissions = null;
-        invitor.Status = OrganizationUserStatusType.Confirmed;
-        var organizationRepository = sutProvider.GetDependency<IOrganizationRepository>();
-        var organizationUserRepository = sutProvider.GetDependency<IOrganizationUserRepository>();
-        var currentContext = sutProvider.GetDependency<ICurrentContext>();
-
-        organizationRepository.GetByIdAsync(organization.Id).Returns(organization);
-        organizationUserRepository.GetManyByOrganizationAsync(organization.Id, OrganizationUserType.Owner)
-            .Returns(new[] { invitor });
-        currentContext.OrganizationOwner(organization.Id).Returns(true);
-        currentContext.ManageUsers(organization.Id).Returns(true);
-
-        await sutProvider.Sut.InviteUsersAsync(organization.Id, invitor.UserId, new (OrganizationUserInvite, string)[] { (invite, null) });
-    }
-
-    [Theory]
-    [OrganizationInviteCustomize(
-        InviteeUserType = OrganizationUserType.User,
-        InvitorUserType = OrganizationUserType.Custom
-    ), BitAutoData]
-    public async Task InviteUser_Passes(Organization organization, IEnumerable<(OrganizationUserInvite invite, string externalId)> invites,
-        OrganizationUser invitor,
-        [OrganizationUser(OrganizationUserStatusType.Confirmed, OrganizationUserType.Owner)] OrganizationUser owner,
-        SutProvider<OrganizationService> sutProvider)
-    {
-        invitor.Permissions = JsonSerializer.Serialize(new Permissions() { ManageUsers = true },
-            new JsonSerializerOptions
-            {
-                PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
-            });
-
-        var organizationRepository = sutProvider.GetDependency<IOrganizationRepository>();
-        var organizationUserRepository = sutProvider.GetDependency<IOrganizationUserRepository>();
-        var currentContext = sutProvider.GetDependency<ICurrentContext>();
-
-        organizationRepository.GetByIdAsync(organization.Id).Returns(organization);
-        organizationUserRepository.GetManyByOrganizationAsync(organization.Id, OrganizationUserType.Owner)
-            .Returns(new[] { owner });
-        currentContext.ManageUsers(organization.Id).Returns(true);
-        currentContext.AccessReports(organization.Id).Returns(true);
-        currentContext.ManageGroups(organization.Id).Returns(true);
-        currentContext.ManagePolicies(organization.Id).Returns(true);
-        currentContext.ManageScim(organization.Id).Returns(true);
-        currentContext.ManageSso(organization.Id).Returns(true);
-        currentContext.AccessEventLogs(organization.Id).Returns(true);
-        currentContext.AccessImportExport(organization.Id).Returns(true);
-        currentContext.CreateNewCollections(organization.Id).Returns(true);
-        currentContext.DeleteAnyCollection(organization.Id).Returns(true);
-        currentContext.DeleteAssignedCollections(organization.Id).Returns(true);
-        currentContext.EditAnyCollection(organization.Id).Returns(true);
-        currentContext.EditAssignedCollections(organization.Id).Returns(true);
-        currentContext.ManageResetPassword(organization.Id).Returns(true);
-
-        await sutProvider.Sut.InviteUsersAsync(organization.Id, invitor.UserId, invites);
-
-        await sutProvider.GetDependency<IMailService>().Received(1)
-            .BulkSendOrganizationInviteEmailAsync(organization.Name,
-                Arg.Is<IEnumerable<(OrganizationUser, ExpiringToken)>>(v => v.Count() == invites.SelectMany(i => i.invite.Emails).Count()), organization.PlanType == PlanType.Free);
-
-        await sutProvider.GetDependency<IEventService>().Received(1).LogOrganizationUserEventsAsync(Arg.Any<IEnumerable<(OrganizationUser, EventType, DateTime?)>>());
-    }
-
-    [Theory]
-    [OrganizationInviteCustomize(
-        InviteeUserType = OrganizationUserType.User,
-        InvitorUserType = OrganizationUserType.Custom
-    ), BitAutoData]
-    public async Task InviteUser_WithEventSystemUser_Passes(Organization organization, EventSystemUser eventSystemUser, IEnumerable<(OrganizationUserInvite invite, string externalId)> invites,
-        OrganizationUser invitor,
-        [OrganizationUser(OrganizationUserStatusType.Confirmed, OrganizationUserType.Owner)] OrganizationUser owner,
-        SutProvider<OrganizationService> sutProvider)
-    {
-        invitor.Permissions = JsonSerializer.Serialize(new Permissions() { ManageUsers = true },
-            new JsonSerializerOptions
-            {
-                PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
-            });
-
-        var organizationRepository = sutProvider.GetDependency<IOrganizationRepository>();
-        var organizationUserRepository = sutProvider.GetDependency<IOrganizationUserRepository>();
-        var currentContext = sutProvider.GetDependency<ICurrentContext>();
-
-        organizationRepository.GetByIdAsync(organization.Id).Returns(organization);
-        organizationUserRepository.GetManyByOrganizationAsync(organization.Id, OrganizationUserType.Owner)
-            .Returns(new[] { owner });
-        currentContext.ManageUsers(organization.Id).Returns(true);
-
-        await sutProvider.Sut.InviteUsersAsync(organization.Id, eventSystemUser, invites);
-
-        await sutProvider.GetDependency<IMailService>().Received(1)
-            .BulkSendOrganizationInviteEmailAsync(organization.Name,
-                Arg.Is<IEnumerable<(OrganizationUser, ExpiringToken)>>(v => v.Count() == invites.SelectMany(i => i.invite.Emails).Count()), organization.PlanType == PlanType.Free);
-
-        await sutProvider.GetDependency<IEventService>().Received(1).LogOrganizationUserEventsAsync(Arg.Any<IEnumerable<(OrganizationUser, EventType, EventSystemUser, DateTime?)>>());
-    }
-
-    [Theory, BitAutoData]
-    public async Task InviteUser_WithSecretsManager_Passes(Organization organization,
-        IEnumerable<(OrganizationUserInvite invite, string externalId)> invites,
-        [OrganizationUser(type: OrganizationUserType.Owner, status: OrganizationUserStatusType.Confirmed)] OrganizationUser savingUser,
-        SutProvider<OrganizationService> sutProvider)
-    {
-        InviteUserHelper_ArrangeValidPermissions(organization, savingUser, sutProvider);
-
-        // Set up some invites to grant access to SM
-        invites.First().invite.AccessSecretsManager = true;
-        var invitedSmUsers = invites.First().invite.Emails.Count();
-
-        // Assume we need to add seats for all invited SM users
-        sutProvider.GetDependency<ICountNewSmSeatsRequiredQuery>()
-            .CountNewSmSeatsRequiredAsync(organization.Id, invitedSmUsers).Returns(invitedSmUsers);
-
-        sutProvider.Sut.InviteUsersAsync(organization.Id, savingUser.Id, invites);
-
-        sutProvider.GetDependency<IUpdateSecretsManagerSubscriptionCommand>().Received(1)
-            .UpdateSubscriptionAsync(Arg.Is<SecretsManagerSubscriptionUpdate>(update =>
-                update.SmSeats == organization.SmSeats + invitedSmUsers &&
-                !update.SmServiceAccountsChanged &&
-                !update.MaxAutoscaleSmSeatsChanged &&
-                !update.MaxAutoscaleSmSeatsChanged));
-    }
-
-    [Theory, BitAutoData]
-    public async Task InviteUser_WithSecretsManager_WhenErrorIsThrown_RevertsAutoscaling(Organization organization,
-        IEnumerable<(OrganizationUserInvite invite, string externalId)> invites,
-        [OrganizationUser(type: OrganizationUserType.Owner, status: OrganizationUserStatusType.Confirmed)] OrganizationUser savingUser,
-        SutProvider<OrganizationService> sutProvider)
-    {
-        var initialSmSeats = organization.SmSeats;
-        InviteUserHelper_ArrangeValidPermissions(organization, savingUser, sutProvider);
-
-        // Set up some invites to grant access to SM
-        invites.First().invite.AccessSecretsManager = true;
-        var invitedSmUsers = invites.First().invite.Emails.Count();
-
-        // Assume we need to add seats for all invited SM users
-        sutProvider.GetDependency<ICountNewSmSeatsRequiredQuery>()
-            .CountNewSmSeatsRequiredAsync(organization.Id, invitedSmUsers).Returns(invitedSmUsers);
-
-        // Mock SecretsManagerSubscriptionUpdateCommand to actually change the organization's subscription in memory
-        sutProvider.GetDependency<IUpdateSecretsManagerSubscriptionCommand>()
-            .UpdateSubscriptionAsync(Arg.Any<SecretsManagerSubscriptionUpdate>())
-            .ReturnsForAnyArgs(Task.FromResult(0)).AndDoes(x => organization.SmSeats += invitedSmUsers);
-
-        // Throw error at the end of the try block
-        sutProvider.GetDependency<IReferenceEventService>().RaiseEventAsync(default).ThrowsForAnyArgs<BadRequestException>();
-
-        sutProvider.Sut.InviteUsersAsync(organization.Id, savingUser.Id, invites);
-
-        // OrgUser is reverted
-        // Note: we don't know what their guids are so comparing length is the best we can do
-        var invitedEmails = invites.SelectMany(i => i.invite.Emails);
-        sutProvider.GetDependency<IOrganizationUserRepository>().Received(1).DeleteManyAsync(
-            Arg.Is<IEnumerable<Guid>>(ids => ids.Count() == invitedEmails.Count()));
-
-        Received.InOrder(() =>
-        {
-            // Initial autoscaling
-            sutProvider.GetDependency<IUpdateSecretsManagerSubscriptionCommand>()
-                .UpdateSubscriptionAsync(Arg.Is<SecretsManagerSubscriptionUpdate>(update =>
-                    update.SmSeats == initialSmSeats + invitedSmUsers &&
-                    !update.SmServiceAccountsChanged &&
-                    !update.MaxAutoscaleSmSeatsChanged &&
-                    !update.MaxAutoscaleSmSeatsChanged));
-
-            // Revert autoscaling
-            sutProvider.GetDependency<IUpdateSecretsManagerSubscriptionCommand>()
-                .UpdateSubscriptionAsync(Arg.Is<SecretsManagerSubscriptionUpdate>(update =>
-                    update.SmSeats == initialSmSeats &&
-                    !update.SmServiceAccountsChanged &&
-                    !update.MaxAutoscaleSmSeatsChanged &&
-                    !update.MaxAutoscaleSmSeatsChanged));
-        });
-    }
-
-    private void InviteUserHelper_ArrangeValidPermissions(Organization organization, OrganizationUser savingUser,
-    SutProvider<OrganizationService> sutProvider)
-    {
-        savingUser.OrganizationId = organization.Id;
-        organization.UseCustomPermissions = true;
-        sutProvider.GetDependency<IOrganizationRepository>().GetByIdAsync(organization.Id).Returns(organization);
-        sutProvider.GetDependency<ICurrentContext>().OrganizationOwner(organization.Id).Returns(true);
-        sutProvider.GetDependency<ICurrentContext>().ManageUsers(organization.Id).Returns(true);
-        sutProvider.GetDependency<IOrganizationUserRepository>().GetManyByOrganizationAsync(savingUser.OrganizationId, OrganizationUserType.Owner)
-            .Returns(new List<OrganizationUser> { savingUser });
     }
 
     [Theory, BitAutoData]
